Rename: com.sh.app -> com.up.and.down
