+++ src/main/java/com/up/and/down/UpProjectApplication.java
@@ -1,13 +1,11 @@
-package com.sh.app;
+package com.up.and.down;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class UpProjectApplication {
-
 	public static void main(String[] args) {
 		SpringApplication.run(UpProjectApplication.class, args);
 	}
-
 }
