+++ src/test/java/com/up/and/down/UpProjectApplicationTests.java
@@ -1,4 +1,4 @@
-package com.sh.app;
+package com.up.and.down;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
